Adding one method to load a file from classpath
I have many dev. environment and it bothers me to change path to load a file. 
Of course, there is a way that I don't change the path but I have to tell the way to my colleagues every time.
So, I just added one method to load a file from classpath and my colleagues can avoid to create the directory in his/her dev. environment and the prod environment as well.

Thanks.
Ax